pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask-dance>=7.1.0",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.1",
    "flask-login>=0.6.3",
    "oauthlib>=3.3.1",
    "pyjwt>=2.10.1",
    "cryptography>=46.0.3",
    "sqlalchemy>=2.0.44",
    "werkzeug>=3.1.3",
    "requests>=2.32.5",
]
